+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+
+- Updated to version 2.0.1 of cockroach-cloud-sdk-go
+
 ## [1.7.7] - 2024-06-20
 
 - Added warning about using private endpoints with old versions.